pyproject.toml
[project]
name = "gemini-rag-lab"
version = "0.1.0"
description = "Basic RAG with LangChain, Streamlit, FastAPI, ChromaDB and Phoenix"
authors = [
    {name = "Gemini Agent"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langchain>=0.3.0,<0.4.0",
    "langchain-community>=0.3.0,<0.4.0",
    "langchain-google-genai>=2.0.0",
    "chromadb>=0.5.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    "streamlit>=1.30.0",
    "arize-phoenix>=4.0.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "langchain-text-splitters>=0.3.0",
    "openinference-instrumentation-langchain (>=0.1.56,<0.2.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic-settings>=2.0.0"
]

[tool.poetry]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"